docs/requirements.txt
breathe==4.11.1
sphinx==1.8.5
GitPython==3.1.0
pydata_sphinx_theme==0.3.0